pack l1 to byte and l1weights to sbyte

+++ Truncator/Evaluate/Weights.cs
@@ -10,7 +10,7 @@ public static class Weights
     public static unsafe short* l0_weight = null;
     public static unsafe short* l0_bias = null;
 
-    public static unsafe short* l1_weight;
+    public static unsafe sbyte* l1_weight;
     public static unsafe float* l1_bias;
 
     public static unsafe float* l2_weight;
@@ -27,7 +27,7 @@ public static class Weights
         l0_weight = (short*)NativeMemory.AlignedAlloc((nuint)sizeof(short) * IN_SIZE * L1_SIZE, 256);
         l0_bias = (short*)NativeMemory.AlignedAlloc((nuint)sizeof(short) * L1_SIZE, 256);
 
-        l1_weight = (short*)NativeMemory.AlignedAlloc((nuint)sizeof(short) * OUT_BUCKETS * L1_SIZE * L2_SIZE, 256);
+        l1_weight = (sbyte*)NativeMemory.AlignedAlloc((nuint)sizeof(sbyte) * OUT_BUCKETS * L1_SIZE * L2_SIZE, 256);
         l1_bias = (float*)NativeMemory.AlignedAlloc((nuint)sizeof(float) * OUT_BUCKETS * L2_SIZE, 256);
 
         l2_weight = (float*)NativeMemory.AlignedAlloc((nuint)sizeof(float) * OUT_BUCKETS * L2_SIZE * L3_SIZE, 256);
@@ -56,7 +56,7 @@ public static class Weights
         for (int buck = 0; buck < OUT_BUCKETS; buck++)
             for (int l2 = 0; l2 < L2_SIZE; l2++)
                 for (int l1 = 0; l1 < L1_SIZE; l1++)
-                    l1_weight[buck * L1_SIZE * L2_SIZE + l2 * L1_SIZE + l1] = (short)net.ReadSByte();
+                    l1_weight[buck * L1_SIZE * L2_SIZE + l2 * L1_SIZE + l1] = net.ReadSByte();
 
         for (int buck = 0; buck < OUT_BUCKETS; buck++)
             for (int l2 = 0; l2 < L2_SIZE; l2++)
